chore: move flag descriptions
